FINAL
Checkpoint 2 v3 is the main file for final checkpoint.
Checkpoint 2 v3 (duplicate v3) is for testing 2 player online gameplay.
To test it, I have hard-coded "mansar" and "mumar" so only send or receive play requests between these users. If you want to try with other users, then the instructions to do that are in the main file (Checkpoint 2 v3).